file handlink not OK

+++ LoraTTNMapper/include/wifi_counter.h
@@ -1,4 +1,4 @@
 #pragma once
 void setup_wifi_counter();
 void wifi_counter_loop();  
-u_int8_t wifi_count_get();
+u_int16_t wifi_count_get();
